+++ src/frontend/packages/cf-autoscaler/package.json
@@ -1,8 +1,12 @@
 {
-  "name": "cf-autoscaler",
+  "name": "@stratosui/cf-autoscaler",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
     "@angular/core": "^6.0.0-rc.0 || ^6.0.0"
+  },
+  "stratos": {
+    "module": "CfAutoscalerPackageModule",
+    "routingModule": "CfAutoscalerRoutingModule"
   }
 }
