+++ src/frontend/packages/cf-autoscaler/src/public_api.ts
@@ -2,5 +2,5 @@
  * Public API Surface of cloud-foundry
  */
 
-// export * from './lib/cloud-foundry.service';
-export * from './lib/cf-autoscaler.module';
+export * from './cf-autoscaler-package.module';
+export * from './cf-autoscaler-routing.module';
