+++ src/frontend/packages/core/src/app.module.ts
@@ -7,7 +7,6 @@ import { Store } from '@ngrx/store';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { debounceTime, filter, withLatestFrom } from 'rxjs/operators';
 
-import { CfAutoscalerModule } from '../../cf-autoscaler/src/cf-autoscaler.module';
 import { SetRecentlyVisitedEntityAction } from '../../store/src/actions/recently-visited.actions';
 import { GeneralEntityAppState, GeneralRequestDataState } from '../../store/src/app-state';
 import { EntityCatalogModule } from '../../store/src/entity-catalog.module';
@@ -116,7 +115,6 @@ class AppStoreDebugModule { }
     AboutModule,
     CustomImportModule,
     XSRFModule,
-    CfAutoscalerModule
   ],
   providers: [
     CustomizationService,
